generate_graphs. py generates variants of a graph pair.
run_simple performs simple GINE graph matching soley based on the edge features.

plot_output then plots the accuracies of these experiments by taking a file with the entire console output of run simple as an input.

For this to work graph_nmr needs to be changed such that source and target features are the same.